# Persistent connection in the keystore