--- a/conf/webpack.debug.config.js
+++ b/conf/webpack.debug.config.js
@@ -13,11 +13,13 @@ module.exports = {
   devtool: 'source-map',
   externals: {
     fs: '{}',
-    mkdirp: '{}',
+    mkdirp: '{}'
   },
   node: {
     console: false,
-    Buffer: true
+    Buffer: true,
+    mkdirp: 'empty',
+    fs: 'empty'
   },
   plugins: [
   ],
@@ -27,8 +29,8 @@ module.exports = {
       path.resolve(__dirname, '../node_modules')
     ],
     alias: {
-      leveldown: 'level-js',
-    },
+      leveldown: 'level-js'
+    }
   },
   resolveLoader: {
     modules: [
@@ -36,5 +38,5 @@ module.exports = {
       path.resolve(__dirname, '../node_modules')
     ],
     moduleExtensions: ['-loader']
-  },
+  }
 }