--- a/conf/webpack.example.config.js
+++ b/conf/webpack.example.config.js
@@ -12,6 +12,8 @@ module.exports = {
   devtool: 'none',
   node: {
     Buffer: true,
+    mkdirp: 'empty',
+    fs: 'empty'
   },
   plugins: [
     new webpack.DefinePlugin({
@@ -24,7 +26,7 @@ module.exports = {
     modules: [
       'node_modules',
       path.resolve(__dirname, '../node_modules')
-    ],
+    ]
   },
   resolveLoader: {
     modules: [
@@ -32,5 +34,5 @@ module.exports = {
       path.resolve(__dirname, '../node_modules')
     ],
     moduleExtensions: ['-loader']
-  },
+  }
 }