--- a/test/counterdb.test.js
+++ b/test/counterdb.test.js
@@ -104,7 +104,7 @@ Object.keys(testAPIs).forEach(API => {
             write: [
               orbitdb1.identity.id,
               orbitdb2.identity.id
-            ],
+            ]
           }
         }