--- a/test/utils/custom-test-cache.js
+++ b/test/utils/custom-test-cache.js
@@ -1,9 +1,7 @@
-const OrbitDbCache = require('orbit-db-cache/Cache.js')
-const localdown = require('localstorage-down')
-
+const OrbitDbCache = require('orbit-db-cache')
 /**
  * A custom cache example. To create a differing custom example, orbitdb cache was
  * used with another abstract-leveldown compliant storage, localdown as an example
  */
 
-module.exports = OrbitDbCache(localdown)
+module.exports = OrbitDbCache