--- a/test/utils/custom-test-keystore.js
+++ b/test/utils/custom-test-keystore.js
@@ -40,7 +40,7 @@ class CustomTestKeystore {
   verify (signature, publicKey, data) {
     return Promise.resolve(true)
   }
-
+    
   getPublic (key) {
     return key.public.marshal()
   }