## Overview
This PR implements the [`orbit-db-storage-adapter`](https://github.com/orbitdb/orbit-db-storage-adapter) paradigm which has two main benefits:

1. Greater performance (see benchmarks below)
2. Easier Interop to redis, mongo, sqlite at the cache and keystore level via the storage adapter

### Benchmarks

Before:
```
$ node benchmarks/benchmark-add.js
Starting IPFS daemon...
95 queries per second, 95 queries in 1 seconds (Oplog: 95)
87 queries per second, 182 queries in 2 seconds (Oplog: 182)
80 queries per second, 262 queries in 3 seconds (Oplog: 262)
88 queries per second, 350 queries in 4 seconds (Oplog: 350)
87 queries per second, 437 queries in 5 seconds (Oplog: 437)
85 queries per second, 522 queries in 6 seconds (Oplog: 522)
83 queries per second, 605 queries in 7 seconds (Oplog: 606)
88 queries per second, 693 queries in 8 seconds (Oplog: 693)
81 queries per second, 774 queries in 9 seconds (Oplog: 774)
--> Average of 86.1 q/s in the last 10 seconds
```

After:
```
$ node benchmarks/benchmark-add.js                             
Starting IPFS daemon...                                        
451 queries per second, 451 queries in 1 seconds (Oplog: 451)  
490 queries per second, 941 queries in 2 seconds (Oplog: 941)  
494 queries per second, 1435 queries in 3 seconds (Oplog: 1435)
513 queries per second, 1948 queries in 4 seconds (Oplog: 1948)
440 queries per second, 2388 queries in 5 seconds (Oplog: 2388)
415 queries per second, 2803 queries in 6 seconds (Oplog: 2803)
416 queries per second, 3219 queries in 7 seconds (Oplog: 3219)
375 queries per second, 3594 queries in 8 seconds (Oplog: 3594)
432 queries per second, 4026 queries in 9 seconds (Oplog: 4026)
--> Average of 445.3 q/s in the last 10 seconds  
```

## Depends On
This PR depends on the following PRs:

- [x] https://github.com/orbitdb/orbit-db-cache/pull/12
- [x] https://github.com/orbitdb/orbit-db-identity-provider/pull/29
- [x] https://github.com/orbitdb/orbit-db-keystore/pull/30
- [x] https://github.com/orbitdb/orbit-db-store/pull/55
- [x] https://github.com/orbitdb/localstorage-level-migration/pull/4
- [x]  https://github.com/orbitdb/ipfs-log/pull/251

## TODO
- [x] Replication still fails intermittently due to underlying js-ipfs bug (still function)
- [x] Deal with failing v0 tests
- [x] Deal with skipped tests
- [x] Update docs