@aphelionz 2019-06-04
@shamb0t @haadcode All tests are passing except _one_ replication test, and it's similar (though not the same) to the issues we're seeing in the v0 tests in `ipfs-log`. We're really close though.

I'm a bit spent today so I'm gonna take a rest but tomorrow I feel like we can together get this taken care of.

Note that I had to remove the usage of `MemStore` from `test/replicate.test.js` because it hadn't been updated to the base32 hashes. We can fix that up and restore it, or remove it completely.